separate egi and rest authentication
